Define interface using typing Protocol
