+++ setup.py
@@ -32,6 +32,7 @@ setuptools.setup(
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
+        'Programming Language :: Python :: 3.9',
         'Development Status :: 4 - Beta',
     ],
     entry_points={
@@ -54,5 +55,6 @@ setuptools.setup(
         'retrying>=1.3.3',
         'Pillow>=8.3',
         'pydicom>=2.2',
+        'typing-extensions>=4.0; python_version < "3.8.0"'
     ]
 )
