+++ src/dicomweb_client/__init__.py
@@ -1,4 +1,4 @@
 __version__ = '0.53.0'
 
-from dicomweb_client.api import DICOMwebClient  # noqa
+from dicomweb_client.api import DICOMwebClient, DICOMfileClient  # noqa
 from dicomweb_client.uri import URI, URISuffix, URIType  # noqa
